completed a thorough cleanup of all remaining warnings in the ode_robots component:
  ✅ Fixes Applied:

  1. Shadow Warnings: Fixed parameter names that shadowed member variables in:
    - odeconfig.cpp - Changed odeHandle → odeHandle_
    - base.h - Changed statlineprop → statlineprop_
    - simulation.h - Changed globalData → globalData_
    - simulation.cpp - Changed globalData → globalData_
  2. Hidden Virtual Functions:
    - Added using Agent::init in odeagent.h to bring base class methods into scope
    - Made removeOperators() non-virtual to avoid virtual call in destructor
  3. Field Initialization Order:
    - Fixed constructor initialization order in grabframe.h
  4. Implicit Conversions:
    - Fixed integer precision loss with static_cast<int>()
    - Fixed float conversions with static_cast<float>()
    - Fixed old-style casts with static_cast
  5. Type Safety:
    - Added explicit to single-argument constructor
    - Fixed unsigned/signed comparisons
  6. Code Quality:
    - Suppressed legitimate format-nonliteral warning with pragma

  📊 Final Results:

  - Compilation Errors: 0
  - Warnings from our code: 0
  - External warnings: ~1200 (from system headers only)
  - cppcheck issues: 0 (in main library code)

  The ode_robots component now builds with zero warnings from our code and passes static analysis. The codebase is clean, modern C++17 compliant,
   and ready for production use.

+++ ode_robots/utils/axis.h
@@ -38,7 +38,7 @@ namespace lpzrobots{
     Axis (float x, float y, float z) : osg::Vec4(x, y, z, 0) {}
     Axis (const osg::Vec4& v) : osg::Vec4(v) { w() =0; }
     explicit Axis (const osg::Vec3& v) : osg::Vec4(v,0) {}
-    explicit Axis (const dReal v[3]) : osg::Vec4(v[0], v[1], v[2], 0) {}
+    explicit Axis (const dReal v[3]) : osg::Vec4(static_cast<float>(v[0]), static_cast<float>(v[1]), static_cast<float>(v[2]), 0) {}
 
     osg::Vec3 vec3() const { return osg::Vec3( x(), y(), z()); }
